requirements.txt
django==1.11
Markdown==2.3.1
PILLOW
South==0.8.1
argparse==1.2.1
boto==2.42.0
crypto==1.1.0
docutils
celery==4.0.2
django-appconf==0.6
django-debug-panel
django-debug-toolbar
django-extensions==1.7.7
django-filter==0.6
django-mptt==0.8.6
#django-robots==3.0
git+https://github.com/Elec/django-robots.git
django-rosetta==0.7.14
djangorestframework==3.7.3
djangorestframework-jwt
factory-boy==2.8.1
ipdb==0.7
ipython==0.13.2
isodate==0.4.9
pyasn1==0.1.7
pycrypto==2.6.1
social-auth-app-django==2.0.0
python-dateutil==2.1
rsa==3.1.2
six==1.6.1
sqlparse==0.1.10
coverage
django_celery_results==1.0.1
celery-with-redis
django-smtp-ssl
django-session-cleanup
django-cors-headers
uWSGI==2.0.14
simplejson==3.10.0
#production
psycopg2

#tests
mock
